requirements.txt
# ─── EURUSD Zero-Loss Trading Bot Dependencies ───────────────────────
# Core Data & Processing
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.26.0
MetaTrader5>=5.0.45       # Live MT5 connection

# Machine Learning Ensemble
scikit-learn>=1.4.0       # RandomForest
xgboost>=2.0.0            # Pattern confluence detection
torch>=2.2.0              # LSTM & FinBERT
transformers>=4.40.0      # NLP Sentiment

# Scrapers & API
playwright>=1.43.0        # TradingView/HTML scraping
tweepy>=4.14.0            # Twitter API (Free Tier)
google-api-python-client  # YouTube Data API v3
beautifulsoup4>=4.12.0    # HTML Parsing
praw>=7.7.1               # Reddit API

# Utilities
requests>=2.31.0
joblib>=1.3.0
python-dotenv>=1.0.0
colorama>=0.4.6

# API & Backend
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9

# Security & Config
cryptography>=42.0.0
pydantic-settings>=2.2.0
aioredis>=2.0.0
